added

+++ lab8/procedures.sql
@@ -1,21 +1,24 @@
-CREATE OR REPLACE PROCEDURE upsert_contact(p_name TEXT, p_phone TEXT)
+CREATE OR REPLACE PROCEDURE upsert_contact(p_name TEXT, p_surname TEXT, p_phone TEXT)
 LANGUAGE plpgsql AS $$
 BEGIN
-    IF EXISTS (SELECT 1 FROM contacts WHERE name = p_name) THEN
+    IF EXISTS (SELECT 1 FROM contacts WHERE name = p_name AND surname = p_surname) THEN
         UPDATE contacts
         SET phone = p_phone
-        WHERE name = p_name;
+        WHERE name = p_name AND surname = p_surname;
     ELSE
-        INSERT INTO contacts(name, phone)
-        VALUES (p_name, p_phone);
+        INSERT INTO contacts(name, surname, phone)
+        VALUES (p_name, p_surname, p_phone);
     END IF;
 END;
 $$;
 
-CREATE OR REPLACE PROCEDURE insert_many_contacts(
+
+CREATE OR REPLACE FUNCTION insert_many_contacts(
     p_names TEXT[],
+    p_surnames TEXT[],
     p_phones TEXT[]
 )
+RETURNS TABLE(invalid_name TEXT, invalid_phone TEXT)
 LANGUAGE plpgsql AS $$
 DECLARE
     i INT;
@@ -24,28 +27,34 @@ BEGIN
     LOOP
         IF p_phones[i] ~ '^\d{11}$' THEN
             
-            IF EXISTS (SELECT 1 FROM contacts WHERE name = p_names[i]) THEN
+            IF EXISTS (
+                SELECT 1 FROM contacts 
+                WHERE name = p_names[i] AND surname = p_surnames[i]
+            ) THEN
                 UPDATE contacts
                 SET phone = p_phones[i]
-                WHERE name = p_names[i];
+                WHERE name = p_names[i] AND surname = p_surnames[i];
             ELSE
-                INSERT INTO contacts(name, phone)
-                VALUES (p_names[i], p_phones[i]);
+                INSERT INTO contacts(name, surname, phone)
+                VALUES (p_names[i], p_surnames[i], p_phones[i]);
             END IF;
 
         ELSE
-            RAISE NOTICE 'Invalid phone: %', p_phones[i];
+            RETURN QUERY SELECT p_names[i], p_phones[i];
         END IF;
     END LOOP;
 END;
 $$;
+
 
 CREATE OR REPLACE PROCEDURE delete_contact(p_name TEXT, p_phone TEXT)
 LANGUAGE plpgsql AS $$
 BEGIN
     IF p_name IS NOT NULL THEN
         DELETE FROM contacts WHERE name = p_name;
-    ELSIF p_phone IS NOT NULL THEN
+    END IF;
+
+    IF p_phone IS NOT NULL THEN
         DELETE FROM contacts WHERE phone = p_phone;
     END IF;
 END;
